Add the filter test cases
