+++ tests/ui-testing/playwright-tests/dashboards/dashboardFilter.spec.js
@@ -1,107 +1,19 @@
 import { test, expect } from "../baseFixtures";
 import logData from "../../cypress/fixtures/log.json";
 import logsdata from "../../../test-data/logs_data.json";
+import { login } from "../../pages/dashLogin.js";
+import { ingestion, removeUTFCharacters } from "../../pages/dashIngestion.js";
+import {
+  waitForDashboardPage,
+  applyQueryButton,
+} from "../../pages/dashCreation.js";
 
 const randomDashboardName =
   "Dashboard_" + Math.random().toString(36).substr(2, 9);
 
 test.describe.configure({ mode: "parallel" });
 
-async function login(page) {
-  await page.goto(process.env["ZO_BASE_URL"], { waitUntil: "networkidle" });
-  // await page.getByText('Login as internal user').click();
-  await page.waitForTimeout(1000);
-  await page
-    .locator('[data-cy="login-user-id"]')
-    .fill(process.env["ZO_ROOT_USER_EMAIL"]);
-
-  // wait for login api response
-  const waitForLogin = page.waitForResponse(
-    (response) =>
-      response.url().includes("/auth/login") && response.status() === 200
-  );
-
-  await page
-    .locator('[data-cy="login-password"]')
-    .fill(process.env["ZO_ROOT_USER_PASSWORD"]);
-  await page.locator('[data-cy="login-sign-in"]').click();
-
-  await waitForLogin;
-
-  await page.waitForURL(process.env["ZO_BASE_URL"] + "/web/", {
-    waitUntil: "networkidle",
-  });
-  await page
-    .locator('[data-test="navbar-organizations-select"]')
-    .getByText("arrow_drop_down")
-    .click();
-  await page.getByRole("option", { name: "default", exact: true }).click();
-}
-
-async function ingestion(page) {
-  const orgId = process.env["ORGNAME"];
-  const streamName = "e2e_automate";
-  const basicAuthCredentials = Buffer.from(
-    `${process.env["ZO_ROOT_USER_EMAIL"]}:${process.env["ZO_ROOT_USER_PASSWORD"]}`
-  ).toString("base64");
-
-  const headers = {
-    Authorization: `Basic ${basicAuthCredentials}`,
-    "Content-Type": "application/json",
-  };
-  const fetchResponse = await fetch(
-    `${process.env.INGESTION_URL}/api/${orgId}/${streamName}/_json`,
-    {
-      method: "POST",
-      headers: headers,
-      body: JSON.stringify(logsdata),
-    }
-  );
-  const response = await fetchResponse.json();
-  console.log(response);
-}
-
-async function waitForDashboardPage(page) {
-  const dashboardListApi = page.waitForResponse(
-    (response) =>
-      /\/api\/.+\/dashboards/.test(response.url()) && response.status() === 200
-  );
-
-  await page.waitForURL(process.env["ZO_BASE_URL"] + "/web/dashboards**");
-
-  await page.waitForSelector(`text="Please wait while loading dashboards..."`, {
-    state: "hidden",
-  });
-  await dashboardListApi;
-  await page.waitForTimeout(500);
-}
-
 test.describe("dashboard UI testcases", () => {
-  // let logData;
-  function removeUTFCharacters(text) {
-    // console.log(text, "tex");
-    // Remove UTF characters using regular expression
-    return text.replace(/[^\x00-\x7F]/g, " ");
-  }
-  async function applyQueryButton(page) {
-    // click on the run query button
-    // Type the value of a variable into an input field
-    const search = page.waitForResponse(logData.applyQuery);
-    await page.waitForTimeout(3000);
-    await page.locator("[data-test='logs-search-bar-refresh-btn']").click({
-      force: true,
-    });
-    // get the data from the search variable
-    await expect.poll(async () => (await search).status()).toBe(200);
-    // await search.hits.FIXME_should("be.an", "array");
-  }
-  // tebefore(async function () {
-  //   // logData("log");
-  //   // const data = page;
-  //   // logData = data;
-
-  //   console.log("--logData--", logData);
-  // });
   test.beforeEach(async ({ page }) => {
     console.log("running before each");
     await login(page);
@@ -109,11 +21,9 @@ test.describe("dashboard UI testcases", () => {
     await ingestion(page);
     await page.waitForTimeout(2000);
 
-    // just to make sure org is set
     const orgNavigation = page.goto(
       `${logData.logsUrl}?org_identifier=${process.env["ORGNAME"]}`
     );
-
     await orgNavigation;
   });
 
@@ -178,19 +88,23 @@ test.describe("dashboard UI testcases", () => {
     //   .locator("i")
     //   .click();
 
-      await page.waitForTimeout(2000);
+    await page.waitForTimeout(2000);
 
-    await page.locator('[data-test="index-dropdown-stream"]').waitFor({ state: 'visible' });
+    await page
+      .locator('[data-test="index-dropdown-stream"]')
+      .waitFor({ state: "visible" });
     await page.locator('[data-test="index-dropdown-stream"]').click();
 
     // await page.getByText('Streamarrow_drop_down').click();
-//     const streamDropdown = page.locator('[data-test="index-dropdown-stream"]').click();
-// await expect(streamDropdown).toBeVisible();
-// await streamDropdown.click();
+    //     const streamDropdown = page.locator('[data-test="index-dropdown-stream"]').click();
+    // await expect(streamDropdown).toBeVisible();
+    // await streamDropdown.click();
 
-     await page.waitForTimeout(2000);
+    await page.waitForTimeout(2000);
 
-    await page.getByRole("option", { name: "e2e_automate", exact: true }).click();
+    await page
+      .getByRole("option", { name: "e2e_automate", exact: true })
+      .click();
 
     // await page.waitForTimeout(2000);
 
@@ -200,20 +114,20 @@ test.describe("dashboard UI testcases", () => {
       )
       .click();
 
-      await page.locator('[data-test="dashboard-apply"]').click();
+    await page.locator('[data-test="dashboard-apply"]').click();
 
     await page.waitForTimeout(3000);
-
 
     // await page
     //   .locator(
     //     '[data-test="field-list-item-logs-e2e_automate-kubernetes_container_name"] [data-test="dashboard-add-filter-data"]'
     //   )
     //   .click();
-    const filterButton = page.locator('[data-test="field-list-item-logs-e2e_automate-kubernetes_container_name"] [data-test="dashboard-add-filter-data"]');
-await expect(filterButton).toBeVisible();
-await filterButton.click();
-
+    const filterButton = page.locator(
+      '[data-test="field-list-item-logs-e2e_automate-kubernetes_container_name"] [data-test="dashboard-add-filter-data"]'
+    );
+    await expect(filterButton).toBeVisible();
+    await filterButton.click();
 
     await page
       .locator('[data-test="dashboard-variable-query-value-selector"]')
@@ -233,13 +147,13 @@ await filterButton.click();
     await page.getByText("Operatorarrow_drop_down").click();
     await page.getByText("=", { exact: true }).click();
     await page.getByLabel("Value").click();
-    await page.getByLabel('Value').fill('$variablename');
+    await page.getByLabel("Value").fill("$variablename");
 
     // await page
     //   .locator("#q-portal--menu--183")
     //   .getByText("variablename")
     //   .click();
-      // await page.getByRole("option", { name: "variablename", exact: true }).click();
+    // await page.getByRole("option", { name: "variablename", exact: true }).click();
 
     await page.locator('[data-test="dashboard-apply"]').click();
 
@@ -259,10 +173,16 @@ await filterButton.click();
     // await page.getByText("arrow_rightQueryAutoPromQLCustom SQL").click();
     await page.waitForTimeout(2000);
 
-    await page.locator('[data-test="dashboard-panel-data-view-query-inspector-btn"]').click();
-  await expect(page.getByRole('cell', { name: 'SELECT histogram(_timestamp) as "x_axis_1", count(_timestamp) as "y_axis_1" FROM "e2e_automate" WHERE kubernetes_container_name = \'ziox\' GROUP BY x_axis_1 ORDER BY x_axis_1 ASC', exact: true })).toBeVisible();
-   
-  
+    await page
+      .locator('[data-test="dashboard-panel-data-view-query-inspector-btn"]')
+      .click();
+    await expect(
+      page.getByRole("cell", {
+        name: 'SELECT histogram(_timestamp) as "x_axis_1", count(_timestamp) as "y_axis_1" FROM "e2e_automate" WHERE kubernetes_container_name = \'ziox\' GROUP BY x_axis_1 ORDER BY x_axis_1 ASC',
+        exact: true,
+      })
+    ).toBeVisible();
+
     await page.locator('[data-test="query-inspector-close-btn"]').click();
 
     await page.locator('[data-test="dashboard-panel-name"]').click();
@@ -282,7 +202,7 @@ await filterButton.click();
       .fill(randomDashboardName);
 
     await page.locator('[data-test="dashboard-add-submit"]').click();
-
+    await page.waitForTimeout(2000);
     await page.locator('[data-test="dashboard-setting-btn"]').click();
     await page.getByRole("tab", { name: "Variables" }).click();
     await page.getByRole("button", { name: "Add Variable" }).click();
@@ -331,19 +251,23 @@ await filterButton.click();
     //   .locator("i")
     //   .click();
 
-      await page.waitForTimeout(2000);
+    await page.waitForTimeout(2000);
 
-    await page.locator('[data-test="index-dropdown-stream"]').waitFor({ state: 'visible' });
+    await page
+      .locator('[data-test="index-dropdown-stream"]')
+      .waitFor({ state: "visible" });
     await page.locator('[data-test="index-dropdown-stream"]').click();
 
     // await page.getByText('Streamarrow_drop_down').click();
-//     const streamDropdown = page.locator('[data-test="index-dropdown-stream"]').click();
-// await expect(streamDropdown).toBeVisible();
-// await streamDropdown.click();
+    //     const streamDropdown = page.locator('[data-test="index-dropdown-stream"]').click();
+    // await expect(streamDropdown).toBeVisible();
+    // await streamDropdown.click();
 
-     await page.waitForTimeout(2000);
+    await page.waitForTimeout(2000);
 
-    await page.getByRole("option", { name: "e2e_automate", exact: true }).click();
+    await page
+      .getByRole("option", { name: "e2e_automate", exact: true })
+      .click();
 
     // await page.waitForTimeout(2000);
 
@@ -352,33 +276,43 @@ await filterButton.click();
         '[data-test="field-list-item-logs-e2e_automate-_timestamp"] [data-test="dashboard-add-y-data"]'
       )
       .click();
-     await page.locator('[data-test="dashboard-apply"]').click();
+    await page.locator('[data-test="dashboard-apply"]').click();
 
-     await page.locator('[data-test="date-time-btn"]').click();
-     await page.locator('[data-test="date-time-relative-6-w-btn"]').click();
-     await page.locator('[data-test="date-time-apply-btn"]').click();
+    await page.locator('[data-test="date-time-btn"]').click();
+    await page.locator('[data-test="date-time-relative-6-w-btn"]').click();
+    await page.locator('[data-test="date-time-apply-btn"]').click();
 
     await page.waitForTimeout(3000);
 
+    // await page
+    //   .locator(
+    //     '[data-test="field-list-item-logs-e2e_automate-kubernetes_container_name"] [data-test="dashboard-add-filter-data"]'
+    //   )
+    //   .click();
+    const filterButton = page.locator(
+      '[data-test="field-list-item-logs-e2e_automate-kubernetes_container_name"] [data-test="dashboard-add-filter-data"]'
+    );
+    await expect(filterButton).toBeVisible();
+    await filterButton.click();
 
     // await page
     //   .locator(
     //     '[data-test="field-list-item-logs-e2e_automate-kubernetes_container_name"] [data-test="dashboard-add-filter-data"]'
     //   )
     //   .click();
-    const filterButton = page.locator('[data-test="field-list-item-logs-e2e_automate-kubernetes_container_name"] [data-test="dashboard-add-filter-data"]');
-await expect(filterButton).toBeVisible();
-await filterButton.click();
 
-    // await page
-    //   .locator(
-    //     '[data-test="field-list-item-logs-e2e_automate-kubernetes_container_name"] [data-test="dashboard-add-filter-data"]'
-    //   )
-    //   .click();
-
-    const filterButton1 = page.locator('[data-test="field-list-item-logs-e2e_automate-kubernetes_container_image"] [data-test="dashboard-add-filter-data"]');
+    const filterButton1 = page.locator(
+      '[data-test="field-list-item-logs-e2e_automate-kubernetes_container_image"] [data-test="dashboard-add-filter-data"]'
+    );
     await expect(filterButton1).toBeVisible();
     await filterButton1.click();
+
+    await page
+      .locator('[data-test="dashboard-variable-query-value-selector"]')
+      .click();
+    await page.getByRole("option", { name: "ziox" }).click();
+
+    await page.locator('[data-test="dashboard-add-condition-add"]').click();
     await page
       .locator(
         '[data-test="dashboard-add-condition-label-0-kubernetes_container_name"]'
@@ -390,7 +324,7 @@ await filterButton.click();
     await page.getByText("Operatorarrow_drop_down").click();
     await page.getByText("=", { exact: true }).click();
     await page.getByLabel("Value").click();
-    await page.getByLabel('Value').fill('$variablename');
+    await page.getByLabel("Value").fill("$variablename");
 
     await page
       .locator(
@@ -400,7 +334,11 @@ await filterButton.click();
     await page
       .locator('[data-test="dashboard-add-condition-condition-1"]')
       .click();
-    await page.getByText("Operatorarrow_drop_down").click();
+    // await page.getByText("Operatorarrow_drop_down").click();
+    await page
+      .locator('[data-test="dashboard-add-condition-operator"]')
+      .click();
+
     await page
       .getByRole("option", { name: "<>" })
       .locator("div")
@@ -411,7 +349,7 @@ await filterButton.click();
       .filter({ hasText: /^Value$/ })
       .nth(2)
       .click();
-      await page.getByLabel('Value').fill('$variablename');
+    await page.getByLabel("Value").fill("$variablename");
 
     await page.locator('[data-test="dashboard-apply"]').click();
     await page.getByText("arrow_rightQueryAutoPromQLCustom SQL").click();
@@ -430,28 +368,26 @@ await filterButton.click();
       name: 'SELECT histogram(_timestamp) as "x_axis_1", count(_timestamp) as "y_axis_1" FROM "e2e_automate" WHERE kubernetes_container_name = \'ziox\' AND kubernetes_container_image <> \'ziox\' GROUP BY x_axis_1 ORDER BY x_axis_1 ASC',
       exact: true,
     });
-    
+
     // Check if the cell is visible
     await expect(cell).toBeVisible();
-    
+
     // Verify it contains the correct text
     await expect(cell).toHaveText(
       'SELECT histogram(_timestamp) as "x_axis_1", count(_timestamp) as "y_axis_1" FROM "e2e_automate" WHERE kubernetes_container_name = \'ziox\' AND kubernetes_container_image <> \'ziox\' GROUP BY x_axis_1 ORDER BY x_axis_1 ASC'
     );
-    
-    await page.locator("#q-portal--dialog--273").getByRole("button").click();
+
+    await page.waitForTimeout(2000);
+    await page.locator('[data-test="query-inspector-close-btn"]').click();
+
     await page.locator('[data-test="dashboard-panel-name"]').click();
     await page.locator('[data-test="dashboard-panel-name"]').fill("test");
     await page.locator('[data-test="dashboard-panel-save"]').click();
   });
 
-  
-  test("Should  apply the  filter group inside group", async ({
-    page,
-  }) => {
-  
+  test("Should  apply the  filter group inside group", async ({ page }) => {
     await page.locator('[data-test="menu-link-\\/dashboards-item"]').click();
-    await waitForDashboardPage(page);           
+    await waitForDashboardPage(page);
     await page.locator('[data-test="dashboard-add"]').click();
     await page.locator('[data-test="add-dashboard-name"]').click();
     await page
@@ -459,6 +395,7 @@ await filterButton.click();
       .fill(randomDashboardName);
 
     await page.locator('[data-test="dashboard-add-submit"]').click();
+    await page.waitForTimeout(2000);
 
     await page.locator('[data-test="dashboard-setting-btn"]').click();
     await page.getByRole("tab", { name: "Variables" }).click();
@@ -501,81 +438,338 @@ await filterButton.click();
 
     await page.waitForTimeout(3000);
     await button.click();
+    await page.waitForTimeout(2000);
 
-    // await page
-    //   .locator("label")
-    //   .filter({ hasText: "Streamarrow_drop_down" })
-    //   .locator("i")
-    //   .click();
-
-      await page.waitForTimeout(2000);
-
-    await page.locator('[data-test="index-dropdown-stream"]').waitFor({ state: 'visible' });
+    await page
+      .locator('[data-test="index-dropdown-stream"]')
+      .waitFor({ state: "visible" });
     await page.locator('[data-test="index-dropdown-stream"]').click();
 
-    // await page.getByText('Streamarrow_drop_down').click();
-//     const streamDropdown = page.locator('[data-test="index-dropdown-stream"]').click();
-// await expect(streamDropdown).toBeVisible();
-// await streamDropdown.click();
+    await page.waitForTimeout(2000);
 
-     await page.waitForTimeout(2000);
-
-    await page.getByRole("option", { name: "e2e_automate", exact: true }).click();
-
-    // await page.waitForTimeout(2000);
+    await page
+      .getByRole("option", { name: "e2e_automate", exact: true })
+      .click();
 
     await page
       .locator(
         '[data-test="field-list-item-logs-e2e_automate-_timestamp"] [data-test="dashboard-add-y-data"]'
       )
       .click();
-     await page.locator('[data-test="dashboard-apply"]').click();
+    await page.locator('[data-test="dashboard-apply"]').click();
 
-     await page.locator('[data-test="date-time-btn"]').click();
-     await page.locator('[data-test="date-time-relative-6-w-btn"]').click();
-     await page.locator('[data-test="date-time-apply-btn"]').click();
+    await page.locator('[data-test="date-time-btn"]').click();
+    await page.locator('[data-test="date-time-relative-6-w-btn"]').click();
+    await page.locator('[data-test="date-time-apply-btn"]').click();
 
     await page.waitForTimeout(3000);
 
+    await page
+      .locator('[data-test="dashboard-variable-query-value-selector"]')
+      .click();
+    await page.getByRole("option", { name: "ziox" }).click();
+
     await page.locator('[data-test="dashboard-add-condition-add"]').click();
-    await page.getByText('Add Group').click();
-    await page.locator('[data-test="dashboard-add-condition-label-0-_timestamp"]').click();
-    await page.getByText('Filters on Fieldarrow_drop_down').click();
-    await page.getByRole('option', { name: 'kubernetes_container_name' }).click();
-    await page.locator('[data-test="dashboard-add-condition-condition-0"]').click();
-    await page.getByText('Operatorarrow_drop_down').click();
-    await page.getByText('=', { exact: true }).click();
-    await page.getByLabel('Value').click();
-    await page.getByLabel('Value').fill('$variablename');
+    await page.getByText("Add Group").click();
+    await page
+      .locator('[data-test="dashboard-add-condition-label-0-_timestamp"]')
+      .click();
+    await page.getByText("Filters on Fieldarrow_drop_down").click();
+    await page
+      .getByRole("option", { name: "kubernetes_container_name" })
+      .click();
+    await page
+      .locator('[data-test="dashboard-add-condition-condition-0"]')
+      .click();
+    await page.getByText("Operatorarrow_drop_down").click();
+    await page.getByText("=", { exact: true }).click();
+    await page.getByLabel("Value").click();
+    await page.getByLabel("Value").fill("$variablename");
 
-    await page.locator('div').filter({ hasText: /^kubernetes_container_namearrow_drop_downcloseadd$/ }).locator('[data-test="dashboard-add-condition-add"]').click();
-    await page.locator('div').filter({ hasText: 'Add Group' }).nth(3).click();
-    await page.locator('[data-test="dashboard-add-condition-label-0-_timestamp"]').click();
-    await page.getByText('Filters on Fieldarrow_drop_down').click();
-    await page.getByRole('option', { name: 'kubernetes_container_image' }).click();
-    await page.locator('[data-test="dashboard-add-condition-condition-0"]').click();
-    await page.getByText('Operatorarrow_drop_down').click();
-    await page.getByRole('option', { name: '<>' }).locator('div').nth(2).click();
+    await page
+      .locator("div")
+      .filter({ hasText: /^kubernetes_container_namearrow_drop_downcloseadd$/ })
+      .locator('[data-test="dashboard-add-condition-add"]')
+      .click();
+    await page.locator("div").filter({ hasText: "Add Group" }).nth(3).click();
+    await page
+      .locator('[data-test="dashboard-add-condition-label-0-_timestamp"]')
+      .click();
+    await page.getByText("Filters on Fieldarrow_drop_down").click();
+    await page
+      .getByRole("option", { name: "kubernetes_container_image" })
+      .click();
+    await page
+      .locator('[data-test="dashboard-add-condition-condition-0"]')
+      .click();
+    await page.getByText("Operatorarrow_drop_down").click();
+    await page
+      .getByRole("option", { name: "<>" })
+      .locator("div")
+      .nth(2)
+      .click();
 
-    await page.getByLabel('Value').click();
-    await page.getByLabel('Value').fill('$variablename');
-    
+    await page.getByLabel("Value").click();
+    await page.getByLabel("Value").fill("$variablename");
+
     await page.locator('[data-test="dashboard-apply"]').click();
-    await page.getByText('<blank>variablenamearrow_drop_down').click();
-    await page.getByRole('option', { name: 'ziox' }).click();
+
     await page.locator('[data-test="dashboard-apply"]').click();
-    await page.getByText('arrow_rightQueryAutoPromQLCustom SQL').click();
-    await expect(page.getByText('\'$variablename\'').first()).toBeVisible();
-    await page.locator('[data-test="dashboard-panel-data-view-query-inspector-btn"]').click();
-    await expect(page.getByRole('cell', { name: 'SELECT histogram(_timestamp) as "x_axis_1", count(_timestamp) as "y_axis_1" FROM "e2e_automate" WHERE (kubernetes_container_name = \'ziox\' AND (kubernetes_container_image <> \'ziox\')) GROUP BY x_axis_1 ORDER BY x_axis_1 ASC', exact: true })).toBeVisible();
-    await page.locator('#q-portal--dialog--104').getByRole('button').click();
+    await page.getByText("arrow_rightQueryAutoPromQLCustom SQL").click();
+    await expect(page.getByText("'$variablename'").first()).toBeVisible();
+    await page
+      .locator('[data-test="dashboard-panel-data-view-query-inspector-btn"]')
+      .click();
+    await expect(
+      page.getByRole("cell", {
+        name: 'SELECT histogram(_timestamp) as "x_axis_1", count(_timestamp) as "y_axis_1" FROM "e2e_automate" WHERE (kubernetes_container_name = \'ziox\' AND (kubernetes_container_image <> \'ziox\')) GROUP BY x_axis_1 ORDER BY x_axis_1 ASC',
+        exact: true,
+      })
+    ).toBeVisible();
+    await page.locator('[data-test="query-inspector-close-btn"]').click();
     await page.locator('[data-test="dashboard-panel-name"]').click();
     await page.locator('[data-test="dashboard-panel-name"]').click();
-    await page.locator('[data-test="dashboard-panel-name"]').fill('test');
+    await page.locator('[data-test="dashboard-panel-name"]').fill("test");
     await page.locator('[data-test="dashboard-panel-save"]').click();
-
-
-    });
   });
 
+  test("Should applly the add group filter with apply the list of value apply successflly ", async ({
+    page,
+  }) => {
+    await page.locator('[data-test="menu-link-\\/dashboards-item"]').click();
+    await waitForDashboardPage(page);
+    await page.locator('[data-test="dashboard-add"]').click();
+    await page.locator('[data-test="add-dashboard-name"]').click();
+    await page
+      .locator('[data-test="add-dashboard-name"]')
+      .fill(randomDashboardName);
 
+    await page.locator('[data-test="dashboard-add-submit"]').click();
+    await page.waitForTimeout(3000);
+
+    await page.locator('[data-test="dashboard-setting-btn"]').click();
+    await page.getByRole("tab", { name: "Variables" }).click();
+    await page.getByRole("button", { name: "Add Variable" }).click();
+    await page
+      .locator("div")
+      .filter({ hasText: /^Name \*$/ })
+      .nth(2)
+      .click();
+
+    await page.getByLabel("Name *").fill("variablename");
+
+    await page
+      .locator("label")
+      .filter({ hasText: "Stream Type *arrow_drop_down" })
+      .locator("i")
+      .click();
+    await page.getByRole("option", { name: "logs" }).click();
+    await page
+      .locator("label")
+      .filter({ hasText: "Stream *arrow_drop_down" })
+      .locator("i")
+      .click();
+    await page.getByRole("option", { name: "e2e_automate" }).click();
+    await page.getByText("Field *arrow_drop_down").click();
+    await page
+      .getByRole("option", { name: "kubernetes_container_name" })
+      .locator("div")
+      .nth(2)
+      .click();
+    await page.getByRole("button", { name: "Save" }).click();
+
+    await page.waitForTimeout(3000);
+
+    const button = page.locator(
+      '[data-test="dashboard-if-no-panel-add-panel-btn"]'
+    );
+    await expect(button).toBeVisible();
+
+    await page.waitForTimeout(3000);
+    await button.click();
+
+    await page.waitForTimeout(2000);
+
+    await page
+      .locator('[data-test="index-dropdown-stream"]')
+      .waitFor({ state: "visible" });
+    await page.locator('[data-test="index-dropdown-stream"]').click();
+
+    await page.waitForTimeout(2000);
+
+    await page
+      .getByRole("option", { name: "e2e_automate", exact: true })
+      .click();
+
+    await page
+      .locator('[data-test="dashboard-x-item-_timestamp-remove"]')
+      .click();
+    await page
+      .locator(
+        '[data-test="field-list-item-logs-e2e_automate-kubernetes_container_name"] [data-test="dashboard-add-x-data"]'
+      )
+      .click();
+    await page
+      .locator(
+        '[data-test="field-list-item-logs-e2e_automate-kubernetes_container_image"] [data-test="dashboard-add-y-data"]'
+      )
+      .click();
+    await page
+      .locator(
+        '[data-test="field-list-item-logs-e2e_automate-kubernetes_namespace_name"] [data-test="dashboard-add-filter-data"]'
+      )
+      .click();
+    await page
+      .locator(
+        '[data-test="dashboard-add-condition-label-0-kubernetes_namespace_name"]'
+      )
+      .click();
+
+    await page.locator('[data-test="dashboard-apply"]').click();
+
+    await page.locator('[data-test="date-time-btn"]').click();
+    await page.locator('[data-test="date-time-relative-6-w-btn"]').click();
+    await page.locator('[data-test="date-time-apply-btn"]').click();
+
+    await page.waitForTimeout(3000);
+
+    await page
+      .locator(
+        '[data-test="dashboard-add-condition-label-0-kubernetes_namespace_name"]'
+      )
+      .click();
+
+    await page
+      .locator('[data-test="dashboard-add-condition-list-tab"]')
+      .waitFor({ state: "visible" });
+    await page
+      .locator('[data-test="dashboard-add-condition-list-tab"]')
+      .click();
+
+    await page.waitForTimeout(2000);
+
+    await page
+      .getByRole("option", { name: "ingress-nginx" })
+      .locator('[data-test="dashboard-add-condition-list-item"]')
+      .click();
+    await page
+      .getByRole("option", { name: "kube-system" })
+      .locator('[data-test="dashboard-add-condition-list-item"]')
+      .click();
+
+    await page.locator('[data-test="dashboard-apply"]').click();
+
+    await page
+      .locator('[data-test="dashboard-panel-data-view-query-inspector-btn"]')
+      .click();
+
+    const cell = await page.getByRole("cell", {
+      name: /SELECT kubernetes_container_name as "x_axis_1", count\(kubernetes_container_image\) as "y_axis_1" FROM "e2e_automate" WHERE kubernetes_namespace_name IN \('ingress-nginx', 'kube-system'\) GROUP BY x_axis_1/,
+    });
+
+    // Ensure the cell is visible
+    await expect(cell.first()).toBeVisible();
+
+    // Verify the text matches
+    await expect(cell.first()).toHaveText(
+      'SELECT kubernetes_container_name as "x_axis_1", count(kubernetes_container_image) as "y_axis_1" FROM "e2e_automate" WHERE kubernetes_namespace_name IN (\'ingress-nginx\', \'kube-system\') GROUP BY x_axis_1'
+    );
+
+    await page.locator('[data-test="query-inspector-close-btn"]').click();
+    await page.locator('[data-test="dashboard-panel-name"]').click();
+    await page.locator('[data-test="dashboard-panel-name"]').click();
+    await page.locator('[data-test="dashboard-panel-name"]').fill("test");
+    await page.locator('[data-test="dashboard-panel-save"]').click();
+  });
+
+  test("Should  apply the  filter using the field button", async ({ page }) => {
+    await page.locator('[data-test="menu-link-\\/dashboards-item"]').click();
+    await waitForDashboardPage(page);
+    await page.locator('[data-test="dashboard-add"]').click();
+    await page.locator('[data-test="add-dashboard-name"]').click();
+    await page
+      .locator('[data-test="add-dashboard-name"]')
+      .fill(randomDashboardName);
+
+    await page.locator('[data-test="dashboard-add-submit"]').click();
+    await page.waitForTimeout(2000);
+
+    await page.locator('[data-test="dashboard-setting-btn"]').click();
+    await page.getByRole("tab", { name: "Variables" }).click();
+    await page.getByRole("button", { name: "Add Variable" }).click();
+    await page
+      .locator("div")
+      .filter({ hasText: /^Name \*$/ })
+      .nth(2)
+      .click();
+
+    // await page.getByLabel('Name *').click();
+    await page.getByLabel("Name *").fill("variablename");
+
+    await page
+      .locator("label")
+      .filter({ hasText: "Stream Type *arrow_drop_down" })
+      .locator("i")
+      .click();
+    await page.getByRole("option", { name: "logs" }).click();
+    await page
+      .locator("label")
+      .filter({ hasText: "Stream *arrow_drop_down" })
+      .locator("i")
+      .click();
+    await page.getByRole("option", { name: "e2e_automate" }).click();
+    await page.getByText("Field *arrow_drop_down").click();
+    await page
+      .getByRole("option", { name: "kubernetes_container_name" })
+      .locator("div")
+      .nth(2)
+      .click();
+    await page.getByRole("button", { name: "Save" }).click();
+
+    await page.waitForTimeout(3000);
+
+    const button = page.locator(
+      '[data-test="dashboard-if-no-panel-add-panel-btn"]'
+    );
+    await expect(button).toBeVisible();
+
+    await page.waitForTimeout(3000);
+    await button.click();
+    await page.waitForTimeout(2000);
+
+    await page
+      .locator('[data-test="index-dropdown-stream"]')
+      .waitFor({ state: "visible" });
+    await page.locator('[data-test="index-dropdown-stream"]').click();
+
+    await page.waitForTimeout(2000);
+
+    await page
+      .getByRole("option", { name: "e2e_automate", exact: true })
+      .click();
+
+    await page
+      .locator(
+        '[data-test="field-list-item-logs-e2e_automate-_timestamp"] [data-test="dashboard-add-y-data"]'
+      )
+      .click();
+    await page.locator('[data-test="dashboard-apply"]').click();
+
+    await page.locator('[data-test="date-time-btn"]').click();
+    await page.locator('[data-test="date-time-relative-6-w-btn"]').click();
+    await page.locator('[data-test="date-time-apply-btn"]').click();
+
+    await page.waitForTimeout(3000);
+
+    await page
+      .locator(
+        '[data-test="field-list-item-logs-e2e_automate-kubernetes_namespace_name"] [data-test="dashboard-add-filter-data"]'
+      )
+      .click();
+    await expect(
+      page.locator(
+        '[data-test="dashboard-add-condition-label-0-kubernetes_namespace_name"]'
+      )
+    ).toBeVisible();
+  });
+});
